+++ CHANGELOG.md
@@ -401,3 +401,8 @@
 ### version 8.3 MARLO BARUA
 
 - Done Implementing Fetch from backend CBSS View
+
+### version 8.4 MARLO BARUA
+
+- Added Sub Categories Chart Function in CBSS USER VIEW
+- Bug fix on Swda Edit (Davao del Oro ==> Davao de Oro)